data loading checkpoint, plot limits fixed, minor comment fixes
The code will now checkpoint after loading all the Elviz CSVs by
looking for a file (data.pkl) that is a pickle of the uber dataframe
and the filename : dataframe dictionary assembled from scanning
CSVs.  On subsequent executions, if the pickle exists, it will be
loaded and the scan of the CSVs will be ignored.  This saves about
2/3 of the data loading time.

The plot limits across all plots are now the same.  The avg. fold
coverage colume occasionally went out to 20k, probably for 16s
pieces so the max avg. fold change value is hard coded to 500x.

There was a minor spealing mistake fix to a comment.